Main

Description

Módulo contendo a realização do jogo Sokoban.

Documentation

type Mapa = ((Int, Int), [(Int, Int)], [[(Int, Int)]], Int, [String], [String], Bool, Int)

Estado do jogo:

Dimensões do mapa

lista de Coordenadas do boneco no mapa

lista de lista de Coordenadas

Score

Mapa após mover

Mapa sem cardinais

Se está ou não completo o nível

lista com os movimentos já feitos

número do mapa

data Command

tipo criado em que U (Up), D(Down), L(Left), R(Right) e Nada(caso não seja nenhum dos casos)

Constructors

U

D

L

R

Nada

tBCPi :: Int

largura e altura do boneco, das caixas e das paredes em Int

tBCPf :: Float

largura e altura do boneco, das caixas e das paredes em Float

main :: IO ()

função principal que invoca o jogo.

loadMapas :: IO [String]

load dos mapas para o jogo

escolheTema :: IO (Picture, Picture, Picture, Picture, Picture)

possibilita o jogador escolher o tema

escolheNivel :: [String] -> IO (String, [String], Int)

possibilita o jogador escolher o nível

defMI :: String -> Int -> Mapa

define o mapa inicial e alguns valores necessários

joga :: mundo -> (mundo -> Picture) -> (Event -> mundo -> mundo) -> IO ()

Função que cria um jogo.

reageTempo :: Float -> mundo -> mundo

Não reage ao passar do tempo.

desenhaMapa :: (Picture, Picture, Picture, Picture, Picture) -> Picture -> Mapa -> Picture

Desenha o jogo dentro da janela

desenhaTab :: [String] -> (Float, Float) -> (Picture, Picture, Picture, Picture) -> Picture

Desenha as caixas, paredes e posiçoes finais

desenhaLinha :: String -> (Float, Float) -> (Picture, Picture, Picture, Picture) -> Picture

Desenha uma linha de caixas, paredes e posiçoes finais

stateGame :: (Int, Int) -> Bool -> Picture

desenha o estado do jogo

reageF :: [String] -> Event -> Mapa -> Mapa

funçao que reage ao teclar do jogador

restart :: Mapa -> Mapa

funçao restart que reinicia o nivel

undo :: Mapa -> Mapa

funçao undo que volta um passo atrás

mcpMapa :: [String] -> Int -> Mapa -> Event -> Mapa

permite passar para o mapa anterior ou seguinte

reageEvento :: Event -> Mapa -> Mapa

Reage ao pressionar das setas do teclado, movendo o boneco 40 pixéis numa direção

moveBoneco :: (Int, Int) -> (Int, Int) -> (Int, Int)

Move o boneco uma coordenada para o lado

convX :: (Int, Int) -> Int -> Int

Converte de Pixeis para coordenada x do mapa em listas de strings

convY :: (Int, Int) -> Int -> Int

Converte de Pixeis para coordenada y do mapa em listas de strings

parteMapa :: [String] -> ([String], [String])

Divide o mapa em tabuleiro e coordenadas. Para isso usamos a função splitAt...

splitAt :: Int -> [a] -> ([a], [a])

splitAt 3 [1,2,3,4,5] == ([1,2,3],[4,5])

contaListasMapa :: [String] -> Int

Devolve o nº de linhas do tabuleiro.

Nesta função usamos isDigit pertencente ao Data.Char para verificarmos se o elemento é ou não um algarismo (0,...9).

sToI :: String -> [Int]

Converte uma String numa lista de dois números.

liToT :: [Int] -> (Int, Int)

Recebe uma lista de dois numeros tuplos.

daCoords :: Int -> [String] -> Int -> [(Int, Int)]

Dá as coordenadas de cada caracter do mapa numa lista.

removeCardinais :: [String] -> [(Int, Int)] -> [String] -> [String]

Remove os cardinais desnecessários do tabuleiro.

removedalinha :: String -> [(Int, Int)] -> [String] -> String

Remove os cardinais desnecessários de uma linha...

devolveCaracterB :: String -> Int -> Int -> Char

Função que devolve o caracter de uma certa lista.

contalinhasB :: [String] -> Int -> Int -> String

Devolve a linha do nº introduzido contando de cima para baixo.

tuploCoord :: [String] -> [(Int, Int)]

Transforma as strings das coordenadas numa lista de tuplos.

mudaCoordsCaixas :: Int -> Int -> [(Int, Int)] -> Command -> [(Int, Int)]

muda as coordenadas das caixas caso o boneco empurre uma delas

colocaCaixa :: [String] -> (Int, Int) -> [String]

Coloca as caixas numa linha.

colocaCaixas :: [String] -> [(Int, Int)] -> [String]

Coloca as caixas no mapa.

remInc :: [String] -> [String]

Função que remove as coordenadas incorretas (as coordenadas que só tem um número ou as coordenadas que têm caracteres inválidos).

verificarCI :: [String] -> Bool

verifica se o mapa tem todas as caixas na posição final

verificalinha :: String -> Bool

verifica se numa linha linha existe alguma caixa fora da posição final.