Main

Description

Módulo contendo a realização da tarefa D que tem como objetivo receber um mapa válido que possui na ultima linha uma lista de comandos, e com essa lista, verificar se ao realizar esses comandos o jogo Sokoban está completo ou incompleto e diz o número de jogadas realizadas.

Funções do mooshak

inStr :: String -> [String]

Parte uma String numa lista de linhas

outStr :: [String] -> String

Junta uma lista de linhas numa String

Funções de teste

correTestes :: IO ()

Corre múltiplos testes para a tarefa

correTeste :: ([String] -> [String]) -> String -> IO ()

Corre um teste para uma tarefa

main :: IO ()

Chama a Tarefa 4

data Command

tipo criado em que U (Up), D(Down), L(Left), R(Right) e Nada(caso não seja nenhum dos casos)

Constructors

U

D

L

R

Nada

Tarefa 4

tarefa4 :: [String] -> [String]

Tarefa onde diz se o jogo está completo ou não e diz o número de jogadas.

Funções da Tarefa 4

parteMapaB :: [String] -> ([String], [String])

Parte mapa em tabuleiro e coordenadas com o comando.

lastM :: [String] -> String

função last

converteString :: String -> [Command]

Converte uma string de comandos numa lista de comandos.

converteC :: Char -> Command

Converte o valor do caracter para o tipo command.

parteCo :: [String] -> String

Devolve a string com a coordenada do boneco inicial.

remVazio :: [String] -> [String]

Remove listas vazias.

verificarCI :: [String] -> Bool

verifica se o mapa tem todas as caixas na posição final

verificalinha :: String -> Bool

verifica se numa linha linha existe alguma caixa fora da posição final.

moveBoneco :: [String] -> (Int, Int) -> [Command] -> [(Int, Int)] -> Int -> ((Int, Int), [(Int, Int)], Int)

movimenta boneco após receber os comandos.

moveB :: [String] -> (Int, Int) -> Command -> [(Int, Int)] -> Int -> ((Int, Int), [(Int, Int)], Int)

Movimenta o boneco ao receber apenas um comando.

verificaMov :: [String] -> (Int, Int) -> [(Int, Int)] -> Command -> Int -> ((Int, Int), [(Int, Int)], Int)

Verifica se o movimento é válido...

mudaCoordsCaixas :: Int -> Int -> [(Int, Int)] -> Command -> [(Int, Int)]

muda as coordenadas das caixas caso o boneco empurre uma delas

colocaCaixa :: [String] -> (Int, Int) -> [String]

Coloca as caixas numa linha.

colocaCaixas :: [String] -> [(Int, Int)] -> [String]

Coloca as caixas no mapa.

devolveCaracterB :: String -> Int -> Int -> Char

Função que devolve o caracter de uma certa lista.

contalinhasB :: [String] -> Int -> Int -> String

Devolve a linha do nº introduzido contando de cima para baixo.

tuploCoord :: [String] -> [(Int, Int)]

Transforma as strings das coordenadas numa lista de tuplos.

contaListasMapa :: [String] -> Int

Devolve o nº de linhas do tabuleiro.

Nesta função usamos isDigit pertencente ao Data.Char para verificarmos se o elemento é ou não um algarismo (0,...9).

sToI :: String -> [Int]

Converte uma String numa lista de dois números.

remInc :: [String] -> [String]

Função que remove as coordenadas incorretas (as coordenadas que só tem um número ou as coordenadas que têm caracteres inválidos).

liToT :: [Int] -> (Int, Int)

Recebe uma lista de dois numeros tuplos.