Main

Description

Módulo contendo a realização de 3 tarefas referentes ao jogo Sokoban na qual, a primeira é referente a verificar se um mapa dado é valido, a segunda referente a simplicar um mapa válido e a terceira referente a calcular o próximo estado do boneco dado um certo comando.

Funções do mooshak

inStr :: String -> [String]

Parte uma String numa lista de linhas

outStr :: [String] -> String

Junta uma lista de linhas numa String

Funções de teste

correTestes :: IO ()

Corre múltiplos testes para as três tarefas

correTeste :: ([String] -> [String]) -> String -> IO ()

Corre um teste para uma tarefa

Funções Tarefas

main :: IO ()

Chama as Tarefas 1, 2 e 3

Tarefa 1

tarefa1 :: [String] -> [String]

Tarefa onde é suposto verificar se um dado mapa é valido ou não.

Funções da Tarefa 1

tarefa1Erros :: [String] -> Int

Devolve o nº da linha que dá erro.

parteMapa :: [String] -> ([String], [String])

Divide o mapa em tabuleiro e coordenadas. Para isso usamos a função splitAt...

splitAt :: Int -> [a] -> ([a], [a])

splitAt 3 [1,2,3,4,5] == ([1,2,3],[4,5])

contaListasMapa :: [String] -> Int

Devolve o nº de linhas do tabuleiro.

Nesta função usamos isDigit pertencente ao Data.Char para verificarmos se o elemento é ou não um algarismo (0,...9).

contaListasBC :: [String] -> Int

Devolve o nº de linhas de coordenadas.

Nesta função usamos isDigit pertencente ao Data.Char para verificarmos se o elemento é ou não um algarismo (0,...9).

validaTab :: Int -> [String] -> Int -> Int

Dá o nº da linha com o erro apos verificar as linhas do tabuleiro

verificalength :: [String] -> Int -> Int

Função que verifica se o tamanho das listas no tabuleiro é sempre igual.

validaLinhaTab :: Int -> String -> Int -> Int

Testa uma linha do tabuleiro.

Nesta função usamos ord que pertence ao Data.Char...

ord :: Char -> Int

Exemplos de utilização:

Se o char a for um # então: ord a == 35 Se o char a for um espaço então: ord a == 32 Se o char a for um . então: ord a == 46

validaCoords :: [String] -> [String] -> Int -> Int

Testa as coordenadas e devolve a linha em caso de haver erros.

validaBoneco :: [(Int, Int)] -> [(Int, Int)] -> Int -> Int

Verifica se o boneco está dentro do mapa.

validaCoordsT :: [String] -> Int -> Int -> Int

Verifica se as coordenadas são válidas, ou seja, se são dois números...

sToI :: String -> [Int]

Converte uma String numa lista de dois números.

remInc :: [String] -> [String]

Função que remove as coordenadas incorretas (as coordenadas que só tem um número ou as coordenadas que têm caracteres inválidos).

liToT :: [Int] -> (Int, Int)

Recebe uma lista de dois numeros tuplos.

devolveCaracter :: String -> Int -> Int -> Char

Devolve o caracter de uma certa lista

contalinhas :: [String] -> Int -> Int -> Int -> String

Devolve a linha correspondente ao número mas contando de baixo para cima...

validaCaixas :: [String] -> [String] -> Int

Verifica a correspondencia entre nº de pontos no tabuleiro e coordenadas.

numeroPontos :: [String] -> Int

Devolve o número de pontos existentes no mapa

validaCaixasR :: [(Int, Int)] -> [(Int, Int)] -> Int -> Int -> Int

verifica se existe ou nao uma caixa sobreposta sobre o boneco ou sobre uma caixa.

juntaErros :: Int -> Int -> Int

Dá o erro com menor

Tarefa 2

tarefa2 :: [String] -> [String]

Tarefa onde é suposto simplicar um mapa válido.

Funções da Tarefa 2

daCoords :: Int -> [String] -> Int -> [(Int, Int)]

Dá as coordenadas de cada caracter do mapa numa lista.

removeCardinais :: [String] -> [(Int, Int)] -> [String] -> [String]

Remove os cardinais desnecessários do tabuleiro.

removedalinha :: String -> [(Int, Int)] -> [String] -> String

Remove os cardinais desnecessários de uma linha...

devolveCaracterB :: String -> Int -> Int -> Char

Função que devolve o caracter de uma certa lista.

contalinhasB :: [String] -> Int -> Int -> String

Devolve a linha do nº introduzido contando de cima para baixo.

tuploCoord :: [String] -> [(Int, Int)]

Transforma as strings das coordenadas numa lista de tuplos.

sndCoord :: [(Int, Int)] -> [Int]

Transforma uma lista de tuplos numa lista com apenas os segundos elementos.

devCoord :: Int -> [(Int, Int)] -> [(Int, Int)]

Devolve os tuplos correspondentes ao número introduzido...

colocaLetras :: [String] -> [String] -> Int -> Int -> [String]

Coloca as letras no tabuleiro.

colocaL :: Int -> String -> [(Int, Int)] -> String

Coloca as letras numa linha..

colocaO :: String -> (Int, Int) -> String

Coloca o sitio do boneco.

colocaIH :: String -> (Int, Int) -> String

Coloca as letras I e H.

auxP :: (String, String) -> String

Insere o I numa lista.

auxE :: (String, String) -> String

Insere o H numa lista.

Auxiliares da Tarefa 3

data Command

tipo criado em que U (Up), D(Down), L(Left), R(Right) e Nada(caso não seja nenhum dos casos)

Constructors

U

D

L

R

Nada

Tarefa 3

tarefa3 :: [String] -> [String]

Tarefa onde é suposto calcular o próximo estado do boneco dado um certo comando.

Funções da Tarefa 3

parteMapaB :: [String] -> ([String], [String])

Parte mapa em tabuleiro e coordenadas com o comando.

converteC :: String -> Command

Converte o valor da string para o tipo command.

parteCo :: [String] -> String

Devolve a string com a coordenada do boneco inicial.

remVazio :: [String] -> [String]

Remove listas vazias.

moveB :: [String] -> String -> Command -> [(Int, Int)] -> [String]

Movimenta o boneco.

verificaMov :: [String] -> (Int, Int) -> Command -> [String]

Verifica se o movimento é válido...

iToS :: Int -> Int -> [String]

Transforma dois números inteiros numa string com um espaço no meio.

colocaCaixa :: [String] -> (Int, Int) -> [String]

Coloca as caixas numa linha.

colocaCaixas :: [String] -> [(Int, Int)] -> [String]

Coloca as caixas no mapa.

tarefa4 :: [String] -> [String]

Tarefa 4