Main

Description

Módulo contendo a realização da tarefa E na qual recebe Pictures e devolve a largura e a altura do menor retangulo envolvendo essas Pictures.

Funções de teste

fTeste :: Picture -> String

tests :: Test

test1 :: Test

test2 :: Test

test3 :: Test

test4 :: Test

test5 :: Test

test6 :: Test

test7 :: Test

test8 :: Test

test9 :: Test

test10 :: Test

test11 :: Test

test12 :: Test

test13 :: Test

test14 :: Test

main :: IO ()

Chama a Tarefa 5

Tarefa 5

tarefa5 :: Picture -> (Float, Float)

Tarefa onde devolve a altura e a largura do menor retangulo envolvendo todas as pictures.

Funções da Tarefa 5

larAlt :: Path -> (Float, Float)

Recebe uma lista de pontos e devolve a largura e a altura a partir dos maximos e minimos de x e y

picLA :: [Picture] -> Path

Transforma uma lista de Pictures em conjunto de pontos

toPath :: Picture -> Path

Transforma as diferentes figuras em conjunto de pontos

translateM :: Float -> Float -> Path -> Path

Faz translate de uma lista de pontos

rotateM :: Float -> Path -> Path

Faz rotate de uma lista de pontos

toRad :: Float -> Float

Transforma graus em radianos

scaleM :: Float -> Float -> Path -> Path

Faz scale de uma lista de pontos